# api-contract.yml
openapi: 3.0.3
info:
  title: Sistema Stock - API Contract
  description: |
    Contrato de API para el sistema de gestión de stock.
    
    ## Arquitectura
    - **Backoffice**: Gestión interna con service_role key
    - **B2B**: Tienda para clientes con auth JWT
    - **OCR**: Microservicio externo en Hetzner
    
    ## Autenticación
    - **Backoffice**: Service role key en headers
    - **B2B**: JWT token de usuario autenticado
    - **OCR**: HTTP Basic Authentication
    
    ## Base URLs
    - **Vercel Apps**: `https://your-app.vercel.app`
    - **OCR Service**: `https://your-ocr.hetzner.com`
  version: 1.0.0
  contact:
    name: Sistema Stock Team
    email: dev@sistema-stock.com
servers:
  - url: https://sistema-stock-b2b.vercel.app/api
    description: B2B Production API
  - url: https://sistema-stock-backoffice.vercel.app/api
    description: Backoffice Production API
  - url: https://ocr-service.hetzner.com
    description: OCR Service (Hetzner)
  - url: http://localhost:3000/api
    description: Development Local
security:
  - BearerAuth: []
  - ServiceRoleAuth: []
  - BasicAuth: []

# =========================================================
# MODELOS DE DATOS
# =========================================================
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token para usuarios B2B autenticados
    
    ServiceRoleAuth:
      type: apiKey
      in: header
      name: Authorization
      description: "Bearer {SUPABASE_SERVICE_ROLE_KEY} para Backoffice"
    
    BasicAuth:
      type: http
      scheme: basic
      description: HTTP Basic para servicio OCR
  schemas:
...
    Order:
      type: object
      properties:
        id:
          type: string
          format: uuid
        customer_id:
          type: string
          format: uuid
        order_number:
          type: string
          example: "20241201-0001"
        status:
          type: string
          enum: [pending, confirmed, prepared, delivered, cancelled]
        total_amount:
          type: number
          format: decimal
          minimum: 0
          example: 156.80
        total_items:
          type: integer
          minimum: 1
          example: 4
        has_backorder:
          type: boolean
          default: false
        backorder_items:
          type: array
          items:
            type: object
        delivery_date:
          type: string
          format: date
        notes:
          type: string
        cancelled_at:
          type: string
          format: date-time
        cancellation_reason:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    OrderItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        order_id:
          type: string
          format: uuid
        product_id:
          type: string
          format: uuid
        product_name:
          type: string
          example: "Tomate Cherry"
        quantity:
          type: number
          format: decimal
          minimum: 0.001
          example: 5.5
        unit_price:
          type: number
          format: decimal
          minimum: 0
          example: 3.80
        discount_percentage:
          type: number
          format: decimal
          minimum: 0
          maximum: 100
          default: 0
        discount_amount:
          type: number
          format: decimal
          minimum: 0
          default: 0
        total_price:
          type: number
          format: decimal
          minimum: 0
          example: 18.05
        created_at:
          type: string
          format: date-time
    CreateOrderRequest:
      type: object
      required:
        - items
        - delivery_date
      properties:
        items:
          type: array
          minItems: 1
          items:
            type: object
            required: [product_id, quantity]
            properties:
              product_id:
                type: string
                format: uuid
              quantity:
                type: number
                format: decimal
                minimum: 0.001
        delivery_date:
          type: string
          format: date
        notes:
          type: string
          maxLength: 500

    # =========================================================
    # CLIENTES Y DESCUENTOS
    # =========================================================
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
          additionalProperties: true
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
    PaginatedResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items: {}
        pagination:
          type: object
          properties:
            page:
              type: integer
              minimum: 1
            limit:
              type: integer
              minimum: 1
              maximum: 100
            total:
              type: integer
              minimum: 0
            total_pages:
              type: integer
              minimum: 0

# =========================================================
# ENDPOINTS API
# =========================================================
paths:
...
  /orders:
    get:
      tags: [B2B - Orders]
      summary: Obtener pedidos del cliente
      description: Lista los pedidos del cliente autenticado con paginación
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          description: Filtrar por estado
          schema:
            type: string
            enum: [pending, confirmed, prepared, delivered, cancelled]
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
      responses:
        '200':
          description: Lista de pedidos
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Order'

    post:
      tags: [B2B - Orders]
      summary: Crear nuevo pedido
      description: |
        Crea un pedido nuevo. El pedido inicia en estado 'pending'.
        Al confirmar, se reserva stock automáticamente.
        Si no hay stock suficiente, se crea backorder.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Pedido creado exitosamente
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Order'
        '400':
          description: Datos inválidos
        '401':
          description: Usuario no autenticado
        '403':
          description: Cliente no aprobado
        '422':
          description: Stock insuficiente
  /orders/{orderId}:
    get:
      tags: [B2B - Orders]
      summary: Obtener detalle de pedido
      security:
        - BearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Detalle del pedido con items
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: '#/components/schemas/Order'
                          - type: object
                            properties:
                              items:
                                type: array
                                items:
                                  $ref: '#/components/schemas/OrderItem'
  /orders/{orderId}/cancel:
    post:
      tags: [B2B - Orders]
      summary: Cancelar pedido
      description: |
        Cancela un pedido. Si estaba confirmado, repone automáticamente el stock reservado.
      security:
        - BearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  description: Motivo de cancelación
      responses:
        '200':
          description: Pedido cancelado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Pedido no se puede cancelar
        '404':
          description: Pedido no encontrado

  # =========================================================
  # BACKOFFICE - ENTRADAS
  # =========================================================
...